+++ apps/assassin/src/features/song/actions.ts
@@ -4,6 +4,7 @@ import { revalidateSeasonBoard } from "@libs/cache/seasonBoard";
 import { getCurrentUser } from "@libs/supabase/auth";
 import SongService from "./service";
 import { SongUpdatePayload } from "./schema";
+import SongRepository from "./repository";
 
 export const getSongAction = async (id: string) => {
   const user = await getCurrentUser();
@@ -32,15 +33,23 @@ export const createSongAction = async (
   return result;
 };
 
-export const updateSongAction = async (id: string, orgId: string, data: SongUpdatePayload) => {
+export const updateSongAction = async (id: string, data: SongUpdatePayload) => {
   const user = await getCurrentUser();
   const result = await SongService.updateSong(id, data, user.id);
+  const orgId = await SongRepository.getSongOrgIdById(id);
+  if (!orgId) {
+    throw new Error(`Song with ID ${id} does not exist.`);
+  }
   revalidateSeasonBoard(orgId);
   return result;
 };
 
-export const deleteSongAction = async (id: string, orgId: string) => {
+export const deleteSongAction = async (id: string) => {
   const user = await getCurrentUser();
+  const orgId = await SongRepository.getSongOrgIdById(id);
+  if (!orgId) {
+    throw new Error(`Song with ID ${id} does not exist.`);
+  }
   const result = await SongService.deleteSong(id, user.id);
   revalidateSeasonBoard(orgId);
   return result;
